chore: move cli logic to dedicated mod
