+++ bin/tempo-zone/Cargo.toml
@@ -14,30 +14,13 @@ workspace = true
 
 [dependencies]
 # tempo
-zone = { path = "../../crates/tempo-zone" }
-zone-primitives.workspace = true
-tempo-chainspec = { workspace = true, features = ["cli"] }
+zone = { path = "../../crates/tempo-zone", features = ["cli"] }
 
 # reth
 reth-cli-util.workspace = true
-reth-consensus.workspace = true
-reth-ethereum = { workspace = true, features = ["cli"] }
-reth-tracing.workspace = true
-
-# alloy
-alloy-primitives.workspace = true
-alloy-signer-local.workspace = true
 
 # tls
 rustls = { version = "0.23", features = ["aws-lc-rs"] }
-
-# misc
-clap.workspace = true
-eyre.workspace = true
-const-hex.workspace = true
-k256.workspace = true
-tokio.workspace = true
-tracing.workspace = true
 
 [[bin]]
 name = "tempo-zone"
@@ -46,4 +29,4 @@ path = "src/main.rs"
 [features]
 default = ["jemalloc"]
 
-jemalloc = ["reth-cli-util/jemalloc", "reth-ethereum/jemalloc"]
+jemalloc = ["reth-cli-util/jemalloc", "zone/jemalloc"]
