+++ crates/tempo-zone/src/lib.rs
@@ -5,6 +5,8 @@
 #![allow(clippy::too_many_arguments)]
 use eyre as _;
 
+#[cfg(feature = "cli")]
+pub mod cli;
 pub mod abi;
 pub mod ext;
 pub use ext::{ChainTempoStateExt, TempoStateExt};
@@ -32,7 +34,7 @@ pub use l1::{
     L1PortalEvents, L1SequencerEvent, L1Subscriber, L1SubscriberConfig,
 };
 pub use l1_state::{PolicyProvider, SharedL1StateCache, SharedPolicyCache};
-pub use node::{ZoneExecutorBuilder, ZoneNode};
+pub use node::{ZoneExecutorBuilder, ZoneNode, ZoneSequencerAddOnsConfig};
 pub use payload::{ZonePayloadAttributes, ZonePayloadTypes};
 pub use withdrawals::{SharedWithdrawalStore, WithdrawalProcessorConfig, WithdrawalStore};
 pub use zonemonitor::{ZoneMonitorConfig, spawn_zone_monitor};
